feat: Update account management features and security measures

+++ src/sdd_cash_manager/schemas/transaction_schema.py
@@ -1,11 +1,12 @@
 from datetime import datetime
+from decimal import Decimal  # New import
 
 from pydantic import BaseModel
 
 
 class BalanceAdjustmentRequest(BaseModel):
     """Request schema for manually adjusting an account balance."""
-    target_balance: float
+    target_balance: Decimal # Changed to Decimal
     adjustment_date: datetime
     description: str
     action_type: str
@@ -15,12 +16,12 @@ class BalanceAdjustmentResponse(BaseModel):
     """Response schema describing a balance adjustment transaction."""
     transaction_id: str
     account_id: str
-    new_balance: float
+    new_balance: Decimal # Changed to Decimal
     id: str
     effective_date: datetime
     booking_date: datetime
     description: str
-    amount: float
+    amount: Decimal # Changed to Decimal
     debit_account_id: str
     credit_account_id: str
     processing_status: str # Use string to match enum value serialization
